requirements.txt
django==3.2.7
django-widget-tweaks
psycopg2
django-dotenv
Faker~=13.15.0
factory_boy